refactor: move collector Event logic to dedicated module
- Extract Event and related structures from context.rs into event.rs
- Update imports and references to match new structure
- Maintain same functionality while improving code organization
